# [Crypto] optimize sha3 in the crypto library