--- a/crypto/hash/sha2.go
+++ b/crypto/hash/sha2.go
@@ -8,17 +8,17 @@ import (
 
 // sha2_256Algo, embeds commonHasher
 type sha2_256Algo struct {
-	*commonHasher
 	hash.Hash
 }
 
 // NewSHA2_256 returns a new instance of SHA2-256 hasher
 func NewSHA2_256() Hasher {
 	return &sha2_256Algo{
-		commonHasher: &commonHasher{
-			algo:       SHA2_256,
-			outputSize: HashLenSha2_256},
 		Hash: sha256.New()}
+}
+
+func (s *sha2_256Algo) Algorithm() HashingAlgorithm {
+	return SHA2_256
 }
 
 // ComputeHash calculates and returns the SHA2-256 output of input byte array.
@@ -37,17 +37,17 @@ func (s *sha2_256Algo) SumHash() Hash {
 
 // sha2_384Algo, embeds commonHasher
 type sha2_384Algo struct {
-	*commonHasher
 	hash.Hash
 }
 
 // NewSHA2_384 returns a new instance of SHA2-384 hasher
 func NewSHA2_384() Hasher {
 	return &sha2_384Algo{
-		commonHasher: &commonHasher{
-			algo:       SHA2_384,
-			outputSize: HashLenSha2_384},
 		Hash: sha512.New384()}
+}
+
+func (s *sha2_384Algo) Algorithm() HashingAlgorithm {
+	return SHA2_384
 }
 
 // ComputeHash calculates and returns the SHA2-384 output of the input byte array.