--- a/crypto/hash/sha2.go
+++ b/crypto/hash/sha2.go
@@ -21,10 +21,11 @@ func (s *sha2_256Algo) Algorithm() HashingAlgorithm {
 	return SHA2_256
 }
 
-// ComputeHash calculates and returns the SHA2-256 output of input byte array.
+// ComputeHash calculates and returns the SHA2-256 digest of the input.
 // It does not reset the state to allow further writing.
 func (s *sha2_256Algo) ComputeHash(data []byte) Hash {
 	s.Reset()
+	// Write does not return an error in this case
 	_, _ = s.Write(data)
 	return s.Sum(nil)
 }
@@ -50,10 +51,11 @@ func (s *sha2_384Algo) Algorithm() HashingAlgorithm {
 	return SHA2_384
 }
 
-// ComputeHash calculates and returns the SHA2-384 output of the input byte array.
+// ComputeHash calculates and returns the SHA2-384 digest of the input.
 // It does not reset the state to allow further writing.
 func (s *sha2_384Algo) ComputeHash(data []byte) Hash {
 	s.Reset()
+	// Write does not return an error in this case
 	_, _ = s.Write(data)
 	return s.Sum(nil)
 }