--- a/crypto/hash/xor_generic.go
+++ b/crypto/hash/xor_generic.go
@@ -24,7 +24,7 @@ var (
 // xorInGeneric xors the bytes in buf into the state; it
 // makes no non-portable assumptions about memory layout
 // or alignment.
-func xorInGeneric(d *state, buf []byte) {
+func xorInGeneric(d *sha3State, buf []byte) {
 	n := len(buf) / 8
 
 	for i := 0; i < n; i++ {
@@ -35,7 +35,7 @@ func xorInGeneric(d *state, buf []byte) {
 }
 
 // copyOutGeneric copies ulint64s to a byte buffer.
-func copyOutGeneric(d *state, b []byte) {
+func copyOutGeneric(b []byte, d *sha3State) {
 	for i := 0; len(b) >= 8; i++ {
 		binary.LittleEndian.PutUint64(b, d.a[i])
 		b = b[8:]