Addresses https://github.com/dapperlabs/flow-go/issues/5342

The Go standard crypto lib implements a generic keccak algorithm that is common for SHA3 (all output versions) and SHAKE. It also provides a flexible API that allows a few features that don't seem to be used by Flow (continue writing in a hasher after a hash was computed for instance).
This PR customizes the Go standard implementation to only support SHA3-256 and SHA3-384, and removes the API flexibility to only keep basic functionalities.

This results in copying a lot of the Go standard library, since most of the functions are private. This also means Flow lib will need to maintain its own implementation of sha3 and keep an eye on future SHA3 implementation updates. 

The changes reduce the computation time, the memory usage, and number of allocations:

Before
```
BenchmarkSha3_256-8             767 ns/op             512 B/op          3 allocs/op
BenchmarkSha3_384-8             746 ns/op             544 B/op          3 allocs/op
```

after
```
BenchmarkComputeHash/SHA3_256-8                 510 ns/op              32 B/op          1 allocs/op
BenchmarkComputeHash/SHA3_384-8                 498 ns/op              48 B/op          1 allocs/op
```