@tarakby 2021-06-02
Slowing down on addressing the comments and merging this PR as I would like to get https://github.com/dapperlabs/flow-go/issues/5563 done first. https://github.com/dapperlabs/flow-go/issues/5563 would allow testing non-amd64 code that is introduced by this PR.