@huitseeker 2021-06-09
> The Go standard crypto lib implements a generic keccak algorithm [...] It also provides a flexible API that allows a few features that don't seem to be used by Flow (_continue writing in a hasher after a hash was computed for instance_).

Am I correct in understanding that what we're not using is
-  the ability to manifest a hash at a partial point in a sequence of updates to a hasher?
- rather than the ability to merge in more inptus to a sha3 hasher?