@simonhf 2021-07-13
> This results in copying a lot of the Go standard library, since most of the functions are private.

@tarakby BTW I'm not advocating to implement things like this, and there are definitely pros and cons, but just in case you didn't know, there is a Golang way to access private functions of another package [1].

[1] https://stackoverflow.com/questions/36009327/how-to-access-private-functions-of-another-package-in-golang/36011850